requirements.txt
pip~=25.3
Django~=6.0
asgiref~=3.11.0
sqlparse~=0.5.4
psycopg2-binary~=2.9.11
django-cors-headers~=4.2.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
pytz~=2025.2
python-dotenv~=1.2.1
Pillow~=10.2.0